requirements.txt
nextcord 
virtualvenv
python-dotenv
requests
